Check that api caller has admin role

+++ grails-app/controllers/com/unifina/controller/api/RemoveUsersProductsController.groovy
@@ -14,6 +14,10 @@ class RemoveUsersProductsController {
 			render(status: 400)
 			return
 		}
+		if (!request.apiUser.isAdmin()) {
+			render(status: 401)
+			return
+		}
 		productService.removeUsersProducts(params.username)
 		render(status: 204)
 	}
